Removed project-specific build/serve commands. Removed projectTypeRouter as we'll no longer supporting this after the move to webpack.
